requirements.txt
pyDes==2.0.1
aliyun-python-sdk-core==2.13.20
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.10.1
oss2==2.12.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
tzlocal==2.0.0
urllib3==1.25.8
PyEmail==0.0.1
PyExecJS==1.5.1
beautifulsoup4==4.9.3
